FP-83 Prices show but are for the cheapest airfare available within the period of the search, not directly connected to the flight for which it appears next to.
FP-70 Dates on View Flights page converted from long to Date

+++ src/main/java/com/FlightPub/model/Flight.java
@@ -191,7 +191,7 @@ public class Flight {
 
     public static Date longToDate(Long in) {
         try{
-            return new Date(in.longValue());
+            return new Date(in);
         } catch (Exception e) {
             System.out.println(e);
             return null;
